% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{assign_stim}
\alias{assign_stim}
\title{Assign stimulus conditions.}
\usage{
assign_stim(object, stim_map)
}
\arguments{
\item{object}{A Seurat object}

\item{stim_map}{A named character vecter, such as \code{c("old" = "new")}.}
}
\description{
Assign stimulus conditions.
}
