% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{calculate_pseudo_bulk}
\alias{calculate_pseudo_bulk}
\title{Calculate pseudo-bulk RNA-Seq data from scRNA-Seq.}
\usage{
calculate_pseudo_bulk(
  sce,
  type = c("counts", "logcounts", "cpm", "vstresiduals"),
  fun = NULL,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object. The following
cell metadata (\code{colData}) columns have to be provided:
\describe{
\item{\code{sample_id}}{unique sample identifiers}
\item{\code{cluster_id}}{subpopulation (cluster) assignments}
\item{\code{group_id}}{experimental group/condition}
}}

\item{type}{Value type of pseudo-bulk data.}

\item{fun}{Function used to calculate pseudobulk, mean or sum. If
\code{fun} is not specified, it will be chosen according \code{type}}

\item{...}{Args passed to \code{muscat::aggregateData}}
}
\value{
A SingleCellExperiment object.
}
\description{
Calculate pseudo-bulk RNA-Seq data from scRNA-Seq.
}
