% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calculateClusterProportion}
\alias{calculateClusterProportion}
\alias{calculateClusterSize}
\title{Calculate the percentage of cell types in samples}
\usage{
calculateClusterProportion(clusters, samples, groups)

calculateClusterSize(clusters, samples, groups)
}
\arguments{
\item{clusters}{A vector or factor which indicates cell types.}

\item{samples}{A vector which indicates sample id.}

\item{groups}{A vector which indicates group id.}
}
\value{
A data frame
}
\description{
Calculate the percentage of cell types in samples
}
\section{Functions}{
\itemize{
\item \code{calculateClusterSize}: Calculate the size of cell
types in samples
}}

