% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_apply}
\alias{diff_state_apply}
\title{Apply funtion on each data.frame of differential state analysis results}
\usage{
diff_state_apply(results, fun, ...)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{...}{Arguments will be passed to \code{fun}}
}
\description{
Apply funtion on each data.frame of differential state analysis results
}
