% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_filter}
\alias{diff_state_filter}
\title{Filter differential state analysis results by FDR and logFC.}
\usage{
diff_state_filter(tbl_list, fdr_limit = 0.05, logfc_limit = 1)
}
\arguments{
\item{tbl_list}{A list of data.frame that are one of results from
\code{\link{pseudobulk_diff_state}}}

\item{fdr_limit}{Upper limit of FDR.}

\item{logfc_limit}{Lower limit of absolute value of logFC.}
}
\value{
Same as \code{tbl_list}
}
\description{
Filter differential state analysis results by FDR and logFC.
}
