% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_format}
\alias{diff_state_format}
\title{Format differential state testing results into an easily filterable table}
\usage{
diff_state_format(results)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}
}
\description{
Format differential state testing results into an easily filterable table
}
