% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_pull}
\alias{diff_state_pull}
\title{Pull out column from diff state results.}
\usage{
diff_state_pull(results, contrasts, clusters, column)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{contrasts}{choose a contrast.}

\item{clusters}{choose a cluster.}

\item{column}{choose a column.}
}
\description{
Pull out column from diff state results.
}
