% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_ranked_lfc}
\alias{diff_state_ranked_lfc}
\title{Prepare ranked logFC vector named with gene names.}
\usage{
diff_state_ranked_lfc(ds, contrasts, clusters)
}
\arguments{
\item{ds}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{contrasts}{choose a contrast.}

\item{clusters}{choose a cluster.}
}
\value{
A ranked logFC vector named gene names
}
\description{
Prepare ranked logFC vector named with gene names.
}
