% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_significant}
\alias{diff_state_significant}
\title{Filter differential state analysis results by FDR and logFC.}
\usage{
diff_state_significant(results, fdr_limit = 0.05, logfc_limit = 1)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{fdr_limit}{Upper limit of FDR.}

\item{logfc_limit}{Lower limit of absolute value of logFC.}
}
\value{
Same as \code{tbl_list}
}
\description{
Filter differential state analysis results by FDR and logFC.
}
