% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_significant_lfc}
\alias{diff_state_significant_lfc}
\title{Prepare significant logFC vector named with gene names.}
\usage{
diff_state_significant_lfc(ds, contrasts, clusters, FDR = 0.05, logFC = 1)
}
\arguments{
\item{ds}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{contrasts}{choose a contrast.}

\item{clusters}{choose a cluster.}

\item{FDR}{Upper limit of FDR.}

\item{logFC}{Lower limit of absolute value of logFC.}
}
\value{
A logFC vector named gene names
}
\description{
Prepare significant logFC vector named with gene names.
}
