% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_subset}
\alias{diff_state_subset}
\title{Subset Differential State Results}
\usage{
diff_state_subset(results, contrasts = NULL, clusters = NULL)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{contrasts}{choose a contrast.}

\item{clusters}{choose a cluster.}
}
\description{
Subset Differential State Results
}
