% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_summary}
\alias{diff_state_summary}
\title{Summarize results from differential state analysis.}
\usage{
diff_state_summary(results)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}
}
\description{
Summarize results from differential state analysis.
}
