% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{dotplot_for_gsea}
\alias{dotplot_for_gsea}
\title{Compare dotplot for GSEA}
\usage{
dotplot_for_gsea(
  object,
  show_category = 5,
  font_size = 12,
  label_format = 30,
  title = ""
)
}
\arguments{
\item{object}{input object}

\item{show_category}{number of enriched terms to display}

\item{font_size}{font size}

\item{label_format}{a numeric value sets wrap length, alternatively
a custom function to format axis labels. by default wraps names
longer that 30 characters}

\item{title}{plot title}
}
\description{
Compare dotplot for GSEA
}
