% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_cnetplot}
\alias{enrich_cnetplot}
\title{Make Gene-Concept Network plot for enrichment analysis}
\usage{
enrich_cnetplot(enr, fold_change = NULL, show_category = 10, circular = FALSE)
}
\arguments{
\item{enr}{A \code{enrichResult} instance.}

\item{fold_change}{A named numeric vector of fold Change.
The names of \code{fold_change} should be consistent with
ID of \code{enr}}

\item{show_category}{How many annotation terms to show.}

\item{circular}{Whether using circular layout.}
}
\description{
Make Gene-Concept Network plot for enrichment analysis
}
