% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_compare_gsea}
\alias{enrich_compare_gsea}
\title{Perform GSEA on gene clusters.}
\usage{
enrich_compare_gsea(ranked_clusters, ...)
}
\arguments{
\item{ranked_clusters}{a list of ranked logFC vector.}

\item{...}{pass to \code{\link[clusterProfiler]{GSEA}}}
}
\value{
A un-official \code{compareClusterResult} object
}
\description{
Perform GSEA on gene clusters.
}
