% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_perform_gsea}
\alias{enrich_perform_gsea}
\title{Perform gene set enrichment analysis}
\usage{
enrich_perform_gsea(ds, go_data, contrasts, clusters, ...)
}
\arguments{
\item{ds}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{go_data}{A data.frame contains four columns:
\enumerate{
\item First column must be named as \code{Gene_ID}
\item Second column must be named as \code{GO_ID}
\item Third column must be named as \code{GO_Name}
\item Fourth column must be named as \code{GO_Level}
}}

\item{contrasts}{choose a contrast.}

\item{clusters}{choose a cluster.}

\item{...}{pass to \code{\link[clusterProfiler]{GSEA}}}
}
\value{
A \code{gseaResult} instance
}
\description{
Perform gene set enrichment analysis
}
