% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_perform_ora}
\alias{enrich_perform_ora}
\title{Perform over representation analysis}
\usage{
enrich_perform_ora(
  ds,
  go_data,
  contrasts,
  clusters,
  FDR = 0.05,
  logFC = 1,
  ...
)
}
\arguments{
\item{ds}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{go_data}{A data.frame contains four columns:
\enumerate{
\item First column must be named as \code{Gene_ID}
\item Second column must be named as \code{GO_ID}
\item Third column must be named as \code{GO_Name}
\item Fourth column must be named as \code{GO_Level}
}}

\item{contrasts}{choose a contrast.}

\item{clusters}{choose a cluster.}

\item{FDR}{Upper limit of FDR.}

\item{logFC}{Lower limit of absolute value of logFC.}

\item{...}{
  Arguments passed on to \code{\link[=enrichment_analysis]{enrichment_analysis}}
  \describe{
    \item{\code{genes}}{A vector of gene id.}
  }}
}
\value{
A \code{enrichResult} instance
}
\description{
Perform over representation analysis
}
