% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Perform enrichment analysis by using clusterProfiler.}
\usage{
enrichment_analysis(genes, go_data)
}
\arguments{
\item{genes}{A vector of gene id.}

\item{go_data}{A data.frame contains four columns:
\enumerate{
\item First column must be named as \code{Gene_ID}
\item Second column must be named as \code{GO_ID}
\item Third column must be named as \code{GO_Name}
\item Fourth column must be named as \code{GO_Level}
}}
}
\description{
Perform enrichment analysis by using clusterProfiler.
}
