% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_cellular_data}
\alias{fetch_cellular_data}
\alias{fetch_cellular_data.character}
\alias{fetch_cellular_data.h5Seurat}
\title{Access Cellular Data from SeuratDisk File}
\usage{
fetch_cellular_data(x, vars, cells = NULL, slot = "data", ...)

\method{fetch_cellular_data}{character}(x, vars, cells = NULL, slot = "data", ...)

\method{fetch_cellular_data}{h5Seurat}(x, vars, ...)
}
\arguments{
\item{x}{h5Seurat filename or connection.}

\item{vars}{List of all variables to fetch.}

\item{cells}{Cells to collect data for (default is all cells)}

\item{slot}{Slot to pull feature data for}

\item{...}{additional parameters}
}
\value{
A data frame with cells as rows and cellular data as columns
}
\description{
Retrieves data (feature expression, PCA scores, metrics, etc.) for a set
of cells in a SeuratDisk file.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Access data from filename.

\item \code{h5Seurat}: Access data from h5Seurat connection.
}}

