% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{methods}
\name{fetch_data}
\alias{fetch_data}
\alias{fetch_data,character-method}
\title{Access Cellular Data from SeuratDisk File

Retrieves data (feature expression, PCA scores, metrics, etc.) for a set
of cells in a SeuratDisk file.}
\usage{
fetch_data(filename, ...)

\S4method{fetch_data}{character}(filename, ...)
}
\arguments{
\item{filename}{h5Seurat filename}

\item{...}{additional parameters}
}
\description{
Access Cellular Data from SeuratDisk File

Retrieves data (feature expression, PCA scores, metrics, etc.) for a set
of cells in a SeuratDisk file.
}
