% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{find_all_avg_expr_genes}
\alias{find_all_avg_expr_genes}
\title{Generate average expression of different sample
for all cell identity classes}
\usage{
find_all_avg_expr_genes(object)
}
\arguments{
\item{object}{A Seurat object.}
}
\value{
A data.frame containing average expression.
}
\description{
Generate average expression of different sample
for all cell identity classes
}
