% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{find_all_diff_expr_genes}
\alias{find_all_diff_expr_genes}
\title{Find differential expressed genes on variable conditions
for each cell cluster.}
\usage{
find_all_diff_expr_genes(object)
}
\arguments{
\item{object}{Seurat object.}
}
\value{
A data frame.
}
\description{
Find differential expressed genes on variable conditions
for each cell cluster.
}
