% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{find_group_marker_genes}
\alias{find_group_marker_genes}
\title{Find group-specific genes.}
\usage{
find_group_marker_genes(
  sce,
  pb,
  target_group,
  method = c("edgeR", "DESeq2", "limma-trend", "limma-voom")
)
}
\arguments{
\item{sce}{A SingleCellExperiment object. The following
cell metadata (\code{colData}) columns have to be provided:
\describe{
\item{\code{sample_id}}{unique sample identifiers}
\item{\code{cluster_id}}{subpopulation (cluster) assignments}
\item{\code{group_id}}{experimental group/condition}
}}

\item{pb}{Object return by \code{\link{calculate_pseudo_bulk}}}

\item{target_group}{Which group to find marker genes.}

\item{method}{Specify which bulk RNA-seq DE methods to apply.}
}
\description{
Find group-specific genes.
}
