% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{findDACombinedClusters}
\alias{findDACombinedClusters}
\title{Combine differential abundant subpopulation}
\usage{
findDACombinedClusters(
  obj_list,
  ref_label,
  resolution = 0.05,
  prune.SNN = 1/15,
  group.singletons = FALSE,
  min.cell = NULL,
  ...
)
}
\arguments{
\item{obj_list}{A list of DAseq results}

\item{ref_label}{The control condition common to all DAseq results}

\item{resolution}{parameter for Seurat function FindClusters(), default 0.05}

\item{prune.SNN}{parameter for Seurat function FindNeighbors(), default 1/15}

\item{group.singletons}{parameter for Seurat function FindClusters(), default True}

\item{min.cell}{integer, number of cells below which a DA region will be removed as outliers, default NULL, use minimum k value in k-vector}

\item{...}{other parameters to pass to Seurat FindClusters()}
}
\description{
This function assumes that each DAseq result is produced
from a different subset of the same integrated cell embedding
and has a common reference condition (aka. control).
}
