% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{findDiffAbundantCells}
\alias{findDiffAbundantCells}
\title{Run DAseq algorithm}
\usage{
findDiffAbundantCells(
  obj,
  ctrl,
  stim,
  reduction = "pca",
  npc = 20,
  pred.thres = NULL,
  k.vector = seq(50, 500, 50),
  resolution = 0.05
)
}
\arguments{
\item{obj}{Seurat object}

\item{ctrl}{Control condition}

\item{stim}{Stimulus condition}

\item{reduction}{Which dimensionality reduction to use}

\item{npc}{How many dimensions to use as input features}

\item{pred.thres}{length-2 vector, top and bottom threshold on DA measure,
default NULL, select significant DA cells based on permutation}

\item{k.vector}{vector, k values to create the score vector}

\item{resolution}{parameter for Seurat function FindClusters(), default 0.05}
}
\description{
This function warp Steps 1~3 of DAseq algorithm from
package \code{\link[DAseq]{getDAcells}}
}
