% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{from_list_to_upset}
\alias{from_list_to_upset}
\title{List of vectors to UpSet data converter}
\usage{
from_list_to_upset(input)
}
\arguments{
\item{input}{List of sets.}
}
\value{
A data.frame contains binary data.
}
\description{
Same to \code{\link[UpSetR]{fromList}}, but add rownames.
}
