% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{generate_seurat_plots}
\alias{generate_seurat_plots}
\title{Generate plots with the Seurat object}
\usage{
generate_seurat_plots(object, output_folder)
}
\arguments{
\item{object}{A seurat object.size.}

\item{output_folder}{Folder to place plots.}
}
\description{
Generate plots with the Seurat object
}
