% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{go_barplot}
\alias{go_barplot}
\title{Make barplot of GO enrichment analysis.}
\usage{
go_barplot(resdf, x = "GeneRatio", show_category = 10, title = "")
}
\arguments{
\item{resdf}{Results returned by \code{enrichment_analysis}}
}
\description{
Make barplot of GO enrichment analysis.
}
