% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_cross_sample}
\alias{heatmap_cross_sample}
\alias{heatmap_cross_sample.Seurat}
\alias{heatmap_cross_sample.SingleCellExperiment}
\alias{heatmap_cross_sample.list}
\alias{heatmap_cross_sample.array}
\title{Plot heatmap across samples.}
\usage{
heatmap_cross_sample(
  x,
  genes,
  samples = NULL,
  clusters = NULL,
  groups = NULL,
  type = c("counts", "logcounts", "cpm", "vstresiduals"),
  ...
)

\method{heatmap_cross_sample}{Seurat}(
  x,
  genes,
  samples = NULL,
  clusters = NULL,
  groups = NULL,
  type = c("counts", "logcounts", "cpm", "vstresiduals"),
  ...
)

\method{heatmap_cross_sample}{SingleCellExperiment}(
  x,
  genes,
  samples = NULL,
  clusters = NULL,
  groups = NULL,
  type = c("counts", "logcounts", "cpm", "vstresiduals"),
  ...
)

\method{heatmap_cross_sample}{list}(
  x,
  genes,
  samples = NULL,
  clusters = NULL,
  groups = NULL,
  type = c("counts", "logcounts", "cpm", "vstresiduals"),
  ...
)

\method{heatmap_cross_sample}{array}(
  x,
  genes,
  samples = NULL,
  clusters = NULL,
  groups = NULL,
  type = c("counts", "logcounts", "cpm", "vstresiduals"),
  ...
)
}
\arguments{
\item{x}{Data object to plot.}

\item{genes}{Genes to be queried.}

\item{samples}{Samples to be queried.}

\item{clusters}{Cell subpopulation to be queried.}

\item{groups}{Group infromation.}

\item{type}{Value type of pseudo-bulk data.}

\item{...}{Args passed to \code{muscat::aggregateData}}
}
\description{
Plot heatmap across samples.
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Plot heatmap for Seurat object.

\item \code{SingleCellExperiment}: Plot heatmap for SingleCellExperiment object.

\item \code{list}: Plot heatmap for a list of matrices.

\item \code{array}: Plot heatmap for array.
}}

