% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{integrated_sample_analysis}
\alias{integrated_sample_analysis}
\title{Perform Seurat integrated analysis for grouped samples}
\usage{
integrated_sample_analysis(obj_list, n_dims = 20, analyze = TRUE, ...)
}
\arguments{
\item{obj_list}{A list of Seurat objects.}

\item{n_dims}{Number of dimensions to use as input.}

\item{analyze}{Perform general Seurat analysis on integrated object or not.}

\item{...}{pass to \code{integration_anchorset}}
}
\value{
A integrated Seurat object.
}
\description{
Perform Seurat integrated analysis for grouped samples
}
