% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{integration_analysis}
\alias{integration_analysis}
\title{Perform general Seurat analysis on integrated object}
\usage{
integration_analysis(object, n_dims = 20, cluster_res = 0.5)
}
\arguments{
\item{object}{Integrated Seurat object}

\item{n_dims}{Number of dimensions to use as input.}

\item{cluster_res}{Resolution for \code{\link[Seurat]{FindClusters}}}
}
\value{
A Seurat object
}
\description{
Perform general Seurat analysis on integrated object
}
