% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{integration_anchorset}
\alias{integration_anchorset}
\title{Get integration anchorset from Seurat objects.}
\usage{
integration_anchorset(
  obj_list,
  n_dims = 20,
  reduction = c("cca", "rpca"),
  k.anchor = 5,
  reference = NULL,
  normalization = "LogNormalize",
  ...
)
}
\arguments{
\item{obj_list}{A list of Seurat objects.}

\item{n_dims}{Number of dimensions to use as input.}

\item{reduction}{Dimensional reduction to perform when finding anchors. Can
be one of:
\itemize{
  \item{cca: Canonical correlation analysis}
  \item{rpca: Reciprocal PCA}
  \item{rlsi: Reciprocal LSI}
}}

\item{k.anchor}{How many neighbors (k) to use when picking anchors}

\item{reference}{A vector specifying the object/s to be used as a reference
during integration. If NULL (default), all pairwise anchors are found (no
reference/s). If not NULL, the corresponding objects in \code{object.list}
will be used as references. When using a set of specified references, anchors
are first found between each query and each reference. The references are
then integrated through pairwise integration. Each query is then mapped to
the integrated reference.}

\item{normalization}{Normalization method.}
}
\value{
Returns an \code{AnchorSet} object that can be used as input
to \code{\link[Seurat]{IntegrateData}}.
}
\description{
Get integration anchorset from Seurat objects.
}
