% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{intersection_counts}
\alias{intersection_counts}
\title{Query element counts of intersections.}
\usage{
intersection_counts(bin, ...)
}
\arguments{
\item{bin}{Binary data produced by \code{from_list_to_upset}}
}
\description{
Query element counts of intersections.
}
