% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{key_mapping}
\alias{key_mapping}
\title{Mapping key according to a given dict.}
\usage{
key_mapping(x, dict)
}
\arguments{
\item{x}{A character vector to convert.}

\item{dict}{A named character defined mapping rule.}
}
\value{
A character vector.
}
\description{
Mapping key according to a given dict.
}
