% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_load.R
\name{load_manifest}
\alias{load_manifest}
\title{Load project manifest from root folder.}
\usage{
load_manifest(base_folder)
}
\arguments{
\item{base_folder}{Project results root folder.}
}
\value{
A list converted from json file.
}
\description{
Load project manifest from root folder.
}
