% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{main_server}
\alias{main_server}
\title{Main server logic}
\usage{
main_server(input, output, session)
}
\arguments{
\item{input, output, session}{Arguments passed by shiny}
}
\description{
Main server logic
}
