% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedobulk.R
\name{make_integrated_psuedo_bulk}
\alias{make_integrated_psuedo_bulk}
\title{Make pseudo-bulk RNA-Seq data from a Seurat object.}
\usage{
make_integrated_psuedo_bulk(
  seurat_object,
  type = c("counts", "logcounts", "cpm", "vstresiduals")
)
}
\arguments{
\item{seurat_object}{Seurat object after integrated analysis.
The \code{sample} and \code{group} infromation must exist in
\code{seurat_object@meta.data} slot.}

\item{type}{Value type of pseudo-bulk data.}
}
\value{
An array with three dimensions. The three dimensions represent
genes (rownames), samples (colnames), and cell subpopulation, respectively.
}
\description{
Make pseudo-bulk RNA-Seq data from a Seurat object.
}
