% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{make_pseudo_bulk}
\alias{make_pseudo_bulk}
\title{Make pseudo-bulk RNA-Seq from scRNA-Seq.}
\usage{
make_pseudo_bulk(
  data_folders,
  normalization = "LogNormalize",
  method = "avg",
  ...
)
}
\arguments{
\item{data_folders}{A vector of Rhapsody WTA results folder path.}

\item{normalization}{Normalization methods to apply. Available values
are same as \code{Seurat::NormalizeData} plus "SCTransform".}

\item{method}{Pseudo-bulk method to apply. Available values
are \code{avg} and \code{sum} .}

\item{...}{pass to normalization functions.}
}
\value{
A data.frame containing pseudo-bulk RNA-Seq expression data.
}
\description{
Make pseudo-bulk RNA-Seq from scRNA-Seq.
}
