% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedobulk.R
\name{make_psuedo_bulk}
\alias{make_psuedo_bulk}
\title{Make psuedo-bulk RNA-Seq from scRNA-Seq.}
\usage{
make_psuedo_bulk(data_folders, normalization = "LogNormalize", method = "avg")
}
\arguments{
\item{data_folders}{A vector of Rhapsody WTA results folder path.}

\item{normalization}{Normalization methods to apply. Available values
are same as \code{Seurat::NormalizeData} plus "SCTransform".}

\item{method}{Psuedo-bulk method to apply. Available values
are \code{avg} and \code{sum} .}
}
\value{
A data.frame containing psuedo-bulk RNA-Seq expression data.
}
\description{
Make psuedo-bulk RNA-Seq from scRNA-Seq.
}
