% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{merge_sample}
\alias{merge_sample}
\title{Merge Seurat objects into one without batch effect correction}
\usage{
merge_sample(obj_list, normalization = "LogNormalize", analyze = TRUE)
}
\arguments{
\item{obj_list}{A list of Seurat objects.}

\item{normalization}{Normalization method}

\item{analyze}{Perform general Seurat analysis on integrated object or not.}
}
\value{
A merged Seurat object
}
\description{
Merge Seurat objects into one without batch effect correction
}
