% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{onlyDAscore}
\alias{onlyDAscore}
\title{Calculate DA Score}
\usage{
onlyDAscore(cell.labels, cell.idx, labels.1, labels.2)
}
\arguments{
\item{cell.labels}{Sample label for all cells.}

\item{cell.idx}{Location of cells belong to a given identity.}

\item{labels.1}{label name(s) that represent condition 1}

\item{labels.2}{label name(s) that represent condition 2}
}
\description{
Calculate DA Score
}
