% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{perform_diff_gene}
\alias{perform_diff_gene}
\title{Perform differentially expressed genes analysis.}
\usage{
perform_diff_gene(object, output_folder, draw_plot = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{output_folder}{Folder to place outputs}

\item{draw_plot}{Draw plots or not.}
}
\description{
Perform differentially expressed genes analysis.
}
