% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{perform_find_all_markers}
\alias{perform_find_all_markers}
\title{Find gene markers for all cell identity classes}
\usage{
perform_find_all_markers(object, output_folder)
}
\arguments{
\item{object}{A Seurat object.}

\item{output_folder}{Folder to place outputs.}
}
\description{
Find gene markers for all cell identity classes
}
