% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{plot_avg_expr_genes}
\alias{plot_avg_expr_genes}
\title{Plot scatter diagram to compare average expression value of genes on
variable conditions for each cell cluster.}
\usage{
plot_avg_expr_genes(avg_genes, diff_genes)
}
\arguments{
\item{avg_genes}{A data frame produced by \code{find_all_avg_expr_genes}}

\item{diff_genes}{A data frame produced by \code{find_all_diff_expr_genes}}
}
\value{
A named list of ggplot2 object.
}
\description{
Plot scatter diagram to compare average expression value of genes on
variable conditions for each cell cluster.
}
