% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDACellScore}
\alias{plotDACellScore}
\title{Plot DA Cell Scores}
\usage{
plotDACellScore(
  obj,
  reduction,
  show_clusters = NULL,
  label = FALSE,
  label_box = FALSE,
  label_cols = NULL,
  label_legend = TRUE,
  circle = FALSE
)
}
\arguments{
\item{obj}{DAseq results}

\item{show_clusters}{Display cell clusters}

\item{label}{Label cell custers}

\item{label_box}{Show label in box}

\item{label_cols}{Label colors}

\item{label_legend}{Show label legend}
}
\value{
ggplot object
}
\description{
The prediction values are overlayed on the 2D embedding
}
