% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDAEmbedding}
\alias{plotDAEmbedding}
\title{Plot DA Embedding}
\usage{
plotDAEmbedding(obj, reduction, group_by = "group")
}
\arguments{
\item{obj}{DAseq results}

\item{reduction}{Cell embeddings to plot}
}
\description{
Plot DA Embedding
}
