% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDAOverlap}
\alias{plotDAOverlap}
\title{Plot the Overlap Between Cell Type and DA Subpopulations}
\usage{
plotDAOverlap(obj)
}
\arguments{
\item{obj}{DAseq results}
}
\description{
Plot the Overlap Between Cell Type and DA Subpopulations
}
