% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDARandomPermutation}
\alias{plotDARandomPermutation}
\title{Plot the Random Permutation of DA Measure}
\usage{
plotDARandomPermutation(obj, size = 0.5)
}
\arguments{
\item{obj}{DAseq results}

\item{size}{Point size}
}
\description{
Plot the Random Permutation of DA Measure
}
