% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDASampleLabel}
\alias{plotDASampleLabel}
\title{Plot DA Sample Labels}
\usage{
plotDASampleLabel(obj, reduction)
}
\arguments{
\item{obj}{DAseq results}
}
\description{
Plot DA Sample Labels
}
