% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDASite}
\alias{plotDASite}
\title{Selected DA cells are highlighted in the 2D embedding}
\usage{
plotDASite(obj, reduction)
}
\arguments{
\item{obj}{DAseq results}

\item{reduction}{Cell embeddings to plot}
}
\description{
Selected DA cells are highlighted in the 2D embedding
}
