% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{pseudobulk_diff_state}
\alias{pseudobulk_diff_state}
\title{Perform differential state analysis on pseudo-bulk data.}
\usage{
pseudobulk_diff_state(
  sce,
  pb,
  contrasts = NULL,
  method = c("edgeR", "DESeq2", "limma-trend", "limma-voom"),
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object. The following
cell metadata (\code{colData}) columns have to be provided:
\describe{
\item{\code{sample_id}}{unique sample identifiers}
\item{\code{cluster_id}}{subpopulation (cluster) assignments}
\item{\code{group_id}}{experimental group/condition}
}}

\item{pb}{Object return by \code{\link{calculate_pseudo_bulk}}}

\item{contrasts}{Character vector specifying contrasts,
see \code{\link[limma]{makeContrasts}}. If \code{NULL}, all possible
combination will be tested.}

\item{method}{Specify which bulk RNA-seq DE methods to apply.}

\item{...}{pass to \code{\link[muscat]{pbDS}}}
}
\description{
Perform differential state analysis on pseudo-bulk data.
}
