% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_expression_st}
\alias{read_expression_st}
\title{Each row records counts for cell-gene combinations that have
non-zero RSEC molecule counts. This data does not contain
non-expressing genes.}
\usage{
read_expression_st(x)
}
\description{
Each row records counts for cell-gene combinations that have
non-zero RSEC molecule counts. This data does not contain
non-expressing genes.
}
