% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_molecules_csv}
\alias{read_molecules_csv}
\title{Read raw expression matrix from Rhapsody WTA pipeline}
\usage{
read_molecules_csv(base_dir = ".")
}
\arguments{
\item{base_dir}{Root folder of Rhapsody WTA pipeline results}
}
\value{
An object constructed from \code{\link[Matrix]{Matrix}}
}
\description{
Reads and molecules are counted only if they have passed all
Rhapsody WTA pipeline filters and have been determined to be
from putative cells.
}
