% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_mols_per_cell}
\alias{read_mols_per_cell}
\title{Each row represents the number of molecules in a cell
for each gene in WTA outputs. A cell is identified with
a unique cell index number under Cell_Index. This data
contains partial non-expressing genes.}
\usage{
read_mols_per_cell(x)
}
\description{
Each row represents the number of molecules in a cell
for each gene in WTA outputs. A cell is identified with
a unique cell index number under Cell_Index. This data
contains partial non-expressing genes.
}
