% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_mtx}
\alias{read_mtx}
\title{Read \code{Matrix} data from cached folder.}
\usage{
read_mtx(base_dir = ".")
}
\arguments{
\item{base_dir}{Folder contains cached \code{\link[Matrix]{Matrix}} object.}
}
\value{
An object constructed from \code{\link[Matrix]{Matrix}}
}
\description{
Read \code{Matrix} data from cached folder.
}
