% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_raw_csv}
\alias{read_raw_csv}
\title{Read raw expression matrix from Rhapsody WTA pipeline}
\usage{
read_raw_csv(base_dir = ".")
}
\arguments{
\item{base_dir}{Root folder of Rhapsody WTA pipeline results}
}
\value{
An object constructed from \code{\link[Matrix]{Matrix}}
}
\description{
Read raw expression matrix from Rhapsody WTA pipeline
}
