% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_rhapsody_wta}
\alias{read_rhapsody_wta}
\title{Read expression data from results folder of Rhapsody WTA pipeline.}
\usage{
read_rhapsody_wta(base_dir, use_mtx = FALSE)
}
\arguments{
\item{base_dir}{The path of results folder.}

\item{use_mtx}{Specify whether to use the sparse matrix constructed
by the \code{Matrix} package.}
}
\value{
A matrix whose column names represent cell index,
row names represent genes.
}
\description{
Read expression data from results folder of Rhapsody WTA pipeline.
}
