% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_unfiltered_csv}
\alias{read_unfiltered_csv}
\title{Read unfiltered expression matrix from Rhapsody WTA pipeline}
\usage{
read_unfiltered_csv(base_dir = ".")
}
\arguments{
\item{base_dir}{Root folder of Rhapsody WTA pipeline results}
}
\value{
An object constructed from \code{\link[Matrix]{Matrix}}
}
\description{
Read the file contain unfiltered tables with cell
labels of ≥5 reads.
}
