% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{reassign_sample}
\alias{reassign_sample}
\title{Reassign stimulus conditions.}
\usage{
reassign_sample(object, sample_map)
}
\arguments{
\item{object}{A Seurat object}

\item{sample_map}{A named character vecter, such as \code{c("old" = "new")}.}
}
\description{
Reassign stimulus conditions.
}
