% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{reassign_stim}
\alias{reassign_stim}
\title{Reassign stimulus conditions.}
\usage{
reassign_stim(object, stim_map)
}
\arguments{
\item{object}{A Seurat object}

\item{stim_map}{A named character vecter, such as \code{c("old" = "new")}.}
}
\description{
Reassign stimulus conditions.
}
