% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run Rhapsody WTA Viewer application}
\usage{
run_app(data_folder = ".", host = "127.0.0.1", port = NULL)
}
\arguments{
\item{data_folder}{A folder containing data processing results。}
}
\description{
Run Rhapsody WTA Viewer application
}
