% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_expression_matrix}
\alias{save_expression_matrix}
\title{Save Matrix data to disk.}
\usage{
save_expression_matrix(mtx, mtx_file)
}
\arguments{
\item{mtx}{An \code{\link[Matrix]{Matrix}} object produced
from \code{read_raw_csv}}

\item{mtx_file}{Output filename}
}
\description{
Save Matrix data to disk.
}
