% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_plot}
\alias{save_plot}
\title{Wrapper for \code{ggsave} with more messages.}
\usage{
save_plot(filename, plot, width = 7, height = 7, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
Wrapper for \code{ggsave} with more messages.
}
