% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_to_disk}
\alias{save_to_disk}
\alias{save_to_disk.Seurat}
\title{Save Single-Cell Object to Disk}
\usage{
save_to_disk(obj, filename, ...)

\method{save_to_disk}{Seurat}(obj, filename, ...)
}
\arguments{
\item{obj}{single-cell data object}

\item{filename}{output filename.}

\item{...}{additional parameters}
}
\description{
This function can be used to save single-cell data object to
disk file.
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Save Seurat Object to Disk
}}

