% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_load.R
\name{search_data}
\alias{search_data}
\title{Search Rhapsody WTA outputs for given folders}
\usage{
search_data(raw_result_folders, base_folder)
}
\arguments{
\item{raw_result_folders}{Result folders of Rhapsody WTA pipeline.}

\item{base_folder}{Prefix for result folders.}
}
\value{
A list of results files for each sample.
}
\description{
Search Rhapsody WTA outputs for given folders
}
