% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{upset_dataframe}
\alias{upset_dataframe}
\title{Convert UpSet binary data to tidy data.frame}
\usage{
upset_dataframe(bin)
}
\arguments{
\item{bin}{UpSet binary data produced from \code{export}}
}
\value{
A data.frame
}
\description{
Convert UpSet binary data to tidy data.frame
}
