% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{upset_plot}
\alias{upset_plot}
\title{Plot UpSet diagram from tidy UpSet data.frame}
\usage{
upset_plot(df, n_intersections = 15, margin_left = 1.5)
}
\arguments{
\item{df}{A data.frame produced from \code{upset_dataframe}}

\item{n_intersections}{How many intersections to show.}

\item{margin_left}{Left margin to show sets label}
}
\description{
Plot UpSet diagram from tidy UpSet data.frame
}
