% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{upset_updown_regulated}
\alias{upset_updown_regulated}
\title{Plot up/down-regulated genes in one diagram}
\usage{
upset_updown_regulated(
  up_list,
  down_list,
  n_intersections = 15,
  margin_left = 1.5
)
}
\arguments{
\item{up_list}{A list of up-regulated gene sets}

\item{down_list}{A list of down-regulated gene sets}

\item{n_intersections}{How many intersections to show.}

\item{margin_left}{Left margin to show sets label}
}
\description{
Plot up/down-regulated genes in one diagram
}
