% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{volcano_diff_state}
\alias{volcano_diff_state}
\title{Volcano plot for differential state analysis.}
\usage{
volcano_diff_state(
  x,
  logfc_cut = 1,
  pval_cut = 0.05,
  y_var = "p_adj.loc",
  title = NULL
)
}
\arguments{
\item{x}{A data.frame contains following column:
\describe{
\item{\code{gene}}{gene names}
\item{\code{logFC}}{log2 fold change}
\item{\code{p_adj.loc}}{adjusted p-value}
\item{\code{cluster_id}}{data source}
}}

\item{logfc_cut}{Where to add cutoff-line for logFC.}

\item{pval_cut}{Where to add cutoff-line for P value.}

\item{y_var}{Which column should be used as \code{y}.}

\item{title}{Plot title.}
}
\description{
Volcano plot for differential state analysis.
}
