% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{barplot_cluster_abundance}
\alias{barplot_cluster_abundance}
\alias{barplot_cluster_abundance.data.frame}
\alias{barplot_cluster_abundance.SingleCellExperiment}
\alias{barplot_cluster_abundance.Seurat}
\title{Plot Abundance Comparison Across Sample Clusters}
\usage{
barplot_cluster_abundance(x)

\method{barplot_cluster_abundance}{data.frame}(df)

\method{barplot_cluster_abundance}{SingleCellExperiment}(sce)

\method{barplot_cluster_abundance}{Seurat}(srt)
}
\arguments{
\item{df}{Data frame to plot, must contains columns:
\describe{
\item{\code{frequency}}{the proportion of cells in a cluster in a sample}
\item{\code{sample_id}}{IDs of samples}
\item{\code{cluster_id}}{IDs of clusters}
\item{\code{group_id}}{IDs of sample groups}
}}

\item{sce}{A SingleCellExperiment object. The following
cell metadata (\code{colData}) columns have to be provided:
\describe{
\item{\code{sample_id}}{unique sample identifiers}
\item{\code{cluster_id}}{subpopulation (cluster) assignments}
\item{\code{group_id}}{experimental group/condition}
}}

\item{srt}{Seurat object after integrated analysis.
The \code{sample} and \code{group} infromation must exist in
\code{seurat_object@meta.data} slot.}
}
\value{
A ggplot2 object
}
\description{
Plot Abundance Comparison Across Sample Clusters
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Plot Abundance Comparison Across Sample
Clusters for data.frame

\item \code{SingleCellExperiment}: Plot Abundance Comparison Across Sample
Clusters for SingleCellExperiment

\item \code{Seurat}: Calculate the proportion of cells from
Seurat object.
}}

