% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_filter}
\alias{diff_state_filter}
\title{Filter differential state analysis results.}
\usage{
diff_state_filter(results, ...)
}
\arguments{
\item{results}{Results returned by \code{\link[muscat]{pbDS}} or
\code{\link{pseudobulk_diff_state}}}

\item{...}{Arguments will be passed to \code{\link[dplyr]{filter}}}
}
\value{
Same as \code{tbl_list}
}
\description{
Filter differential state analysis results.
}
