% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_state.R
\name{diff_state_summary}
\alias{diff_state_summary}
\title{Summarize results from differential state analysis.}
\usage{
diff_state_summary(tbl_list, sce)
}
\arguments{
\item{tbl_list}{A list of data.frame that are one of results from
\code{\link{pseudobulk_diff_state}}}

\item{sce}{A SingleCellExperiment object. The following
cell metadata (\code{colData}) columns have to be provided:
\describe{
\item{\code{sample_id}}{unique sample identifiers}
\item{\code{cluster_id}}{subpopulation (cluster) assignments}
\item{\code{group_id}}{experimental group/condition}
}}
}
\description{
Summarize results from differential state analysis.
}
