% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_barplot_by_level}
\alias{enrich_barplot_by_level}
\title{Plot barplot of each GO level.}
\usage{
enrich_barplot_by_level(resdf, x = "GeneRatio", show_category = 20)
}
\arguments{
\item{resdf}{Results returned by \code{enrichment_analysis}}
}
\description{
Plot barplot of each GO level.
}
