% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_data}
\alias{fetch_data}
\alias{fetch_data.character}
\alias{fetch_data.h5Seurat}
\title{Access Cellular Data from SeuratDisk File}
\usage{
fetch_data(obj, ...)

\method{fetch_data}{character}(obj, ...)

\method{fetch_data}{h5Seurat}(obj, ...)
}
\arguments{
\item{obj}{h5Seurat filename or connection}

\item{...}{additional parameters}
}
\description{
Retrieves data (feature expression, PCA scores, metrics, etc.) for a set
of cells in a SeuratDisk file.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Access data from filename.

\item \code{h5Seurat}: Access data from h5Seurat connection.
}}

