% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{find_all_conserved_markers}
\alias{find_all_conserved_markers}
\title{Find conserved markers for all cell identity classes}
\usage{
find_all_conserved_markers(object)
}
\arguments{
\item{object}{A Seurat object.}
}
\value{
A data.frame containing conserved markers.
}
\description{
Find conserved markers for all cell identity classes
}
