% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{findDiffAbundantMarkers}
\alias{findDiffAbundantMarkers}
\title{Find markers for differential abundant subpopulation}
\usage{
findDiffAbundantMarkers(
  obj,
  da,
  method = c("COSG", "Seurat", "STG"),
  top_n = NULL,
  return_raw = FALSE,
  GPU = "",
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{da}{DAseq results}

\item{method}{Which method should be used to find markers.
Method \code{COSG} is based on \code{\link[COSG]{cosg}}, \code{Seruat} is
based on \code{\link[DAseq]{SeuratMarkerFinder}} and \code{STG} is baed
on \code{\link[DAseq]{STGmarkerFinder}} which requires python environment.}

\item{top_n}{Return the top N markers from results. The order of markers
is determined by scores for \code{COSG} and by average log2FC then P value
for \code{Seurat} and \code{STG}.}

\item{return_raw}{Return raw results produced by specified \code{method}.}

\item{GPU}{Which GPU to use (GPU IDs), default using CPU. Note: this value
will be used to set \code{CUDA_VISIBLE_DEVICES} environment.}

\item{...}{Additional arguments passed to marker finder.}
}
\value{
A list of marker genes
}
\description{
Find markers for differential abundant subpopulation
}
