% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{genecount_diff_state}
\alias{genecount_diff_state}
\title{Plot number of genes per subpopulation per contrast.}
\usage{
genecount_diff_state(results, simplify = FALSE)
}
\arguments{
\item{results}{Results from \code{\link{pseudobulk_diff_state}}}

\item{simplify}{If \code{TRUE}, subpopulation without DS genes
won't be shown.}
}
\description{
Plot number of genes per subpopulation per contrast.
}
