% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{genecount_diff_state}
\alias{genecount_diff_state}
\title{Plot number of genes per subpopulation per contrast.}
\usage{
genecount_diff_state(results, fdr_limit = 0.05, logfc_limit = 1)
}
\arguments{
\item{results}{results from \code{\link{pseudobulk_diff_state}}}

\item{fdr_limit}{Upper limit of FDR.}

\item{logfc_limit}{Lower limit of absolute value of logFC.}
}
\description{
Plot number of genes per subpopulation per contrast.
}
