% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{integrated_sample_analysis}
\alias{integrated_sample_analysis}
\title{Perform Seurat integrated analysis for stimulated samples}
\usage{
integrated_sample_analysis(obj_list, dimensionality = 20)
}
\arguments{
\item{obj_list}{A list of Seurat objects.}

\item{dimensionality}{Number of dimensions to use as input.}
}
\value{
A integrated Seurat object.
}
\description{
Perform Seurat integrated analysis for stimulated samples
}
