% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDACellLabel}
\alias{plotDACellLabel}
\title{DA subpopulation clustering result is shown in the 2D embedding}
\usage{
plotDACellLabel(obj, reduction)
}
\arguments{
\item{obj}{DAseq results}
}
\description{
DA subpopulation clustering result is shown in the 2D embedding
}
