% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{plotDACellScore}
\alias{plotDACellScore}
\title{Plot DA Cell Scores}
\usage{
plotDACellScore(
  obj,
  reduction,
  order = FALSE,
  label = FALSE,
  circle = FALSE,
  point_size = 1,
  point_alpha = 1,
  text_size = 4,
  theme_size = 10
)
}
\arguments{
\item{obj}{DAseq results}

\item{reduction}{Cell embeddings to plot}

\item{order}{Order cells by DA score}

\item{label}{Label cell custers}

\item{circle}{Draw circle around clusters}

\item{point_size}{Set point size}

\item{point_alpha}{Set point alpha}

\item{text_size}{Set text size}

\item{theme_size}{Set theme size}
}
\value{
ggplot object
}
\description{
The prediction values are overlayed on the 2D embedding
}
