% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{prepare_muscat_sce}
\alias{prepare_muscat_sce}
\title{Prepare SingleCellExperiment object for muscat package.}
\usage{
prepare_muscat_sce(seurat_object)
}
\arguments{
\item{seurat_object}{Seurat object after integrated analysis.
The \code{sample} and \code{group} infromation must exist in
\code{seurat_object@meta.data} slot.}
}
\value{
A SingleCellExperiment object.
}
\description{
Prepare SingleCellExperiment object for muscat package.
}
