% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{methods}
\name{save_to_disk}
\alias{save_to_disk}
\alias{save_to_disk,Seurat-method}
\title{Save Single-Cell Object to Disk}
\usage{
save_to_disk(obj, filename, ...)

\S4method{save_to_disk}{Seurat}(obj, filename, ...)
}
\arguments{
\item{obj}{single-cell data object}

\item{filename}{output filename.}

\item{...}{additional parameters}
}
\description{
This function can be used to save single-cell data object to
disk file.
}
