% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{upset_plot}
\alias{upset_plot}
\title{Plot UpSet diagram from tidy UpSet data.frame}
\usage{
upset_plot(
  df,
  n_intersections = 15,
  y_expansion = ggplot2::expansion(c(0, 0.1), c(0, 0)),
  plot_margin = ggplot2::margin(0.5, 0.5, 0.5, 1.5, unit = "cm"),
  ...
)
}
\arguments{
\item{df}{A data.frame produced from \code{upset_dataframe}}

\item{n_intersections}{How many intersections to show.}

\item{y_expansion}{Expansion for y-axis, see \code{\link[ggplot2]{expansion}}}

\item{plot_margin}{margin around entire plot (unit with the sizes of the top,
right, bottom, and left margins)}

\item{...}{pass to \code{\link[ggupset]{scale_x_upset}}}
}
\description{
Plot UpSet diagram from tidy UpSet data.frame
}
