% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{upset_updown_regulated}
\alias{upset_updown_regulated}
\title{Plot up/down-regulated genes in one diagram}
\usage{
upset_updown_regulated(
  up_list,
  down_list,
  n_intersections = 15,
  x_expansion = ggplot2::expansion(c(0, 0), c(0.8, 0.8)),
  y_expansion = ggplot2::expansion(c(0, 0.1), c(0, 0)),
  plot_margin = ggplot2::margin(0.5, 0.5, 0.5, 1.5, unit = "cm"),
  ...
)
}
\arguments{
\item{up_list}{A list of up-regulated gene sets}

\item{down_list}{A list of down-regulated gene sets}

\item{n_intersections}{How many intersections to show.}

\item{y_expansion}{Expansion for y-axis, see \code{\link[ggplot2]{expansion}}}

\item{plot_margin}{margin around entire plot (unit with the sizes of the top,
right, bottom, and left margins)}

\item{...}{pass to \code{\link[ggupset]{scale_x_upset}}}
}
\description{
Plot up/down-regulated genes in one diagram
}
